+++ slurm/batch_finalize.sh
@@ -53,6 +53,7 @@ module reset
 module load miniconda
 eval "$(conda shell.bash hook)"
 conda activate corpus
+corpus_check_checkout
 
 ENRICH_BHL="${ENRICH_BHL:-0}"
 SKIP_BUNDLE="${SKIP_BUNDLE:-0}"